Reverted SingleUseTokenProvider TypeScript conversion
no issue

The TypeScript version requires tsx in job runtime contexts where the package is not available, causing scheduled jobs to fail during module loading.
