+++ ghost/core/core/server/services/jobs/job-service.js
@@ -45,11 +45,5 @@ const initTestMode = () => {
 
 const jobManager = new JobManager({errorHandler, workerMessageHandler, JobModel: models.Job, domainEvents, config, events});
 
-// Add Node.js flags for TypeScript support in worker threads
-// This enables --experimental-strip-types for Node < 22.18.0
-jobManager.bree.config.worker = {
-    execArgv: ['--experimental-strip-types', '--import=tsx']
-};
-
 module.exports = jobManager;
 module.exports.initTestMode = initTestMode;
