+++ ghost/core/core/server/services/newsletters/index.js
@@ -1,9 +1,8 @@
 const NewslettersService = require('./newsletters-service.js');
-const {SingleUseTokenProvider} = require('../members/single-use-token-provider');
+const SingleUseTokenProvider = require('../members/single-use-token-provider');
 const mail = require('../mail');
 const models = require('../../models');
 const urlUtils = require('../../../shared/url-utils');
-const settingsCache = require('../../../shared/settings-cache');
 const limitService = require('../limits');
 const labs = require('../../../shared/labs');
 const emailAddressService = require('../email-address');
@@ -20,8 +19,7 @@ module.exports = new NewslettersService({
         SingleUseTokenModel: models.SingleUseToken,
         validityPeriod: MAGIC_LINK_TOKEN_VALIDITY,
         validityPeriodAfterUsage: MAGIC_LINK_TOKEN_VALIDITY_AFTER_USAGE,
-        maxUsageCount: MAGIC_LINK_TOKEN_MAX_USAGE_COUNT,
-        secret: settingsCache.get('members_otc_secret')
+        maxUsageCount: MAGIC_LINK_TOKEN_MAX_USAGE_COUNT
     }),
     urlUtils,
     limitService,
