+++ ghost/core/core/server/services/settings/settings-service.js
@@ -13,7 +13,7 @@ const SettingsBREADService = require('./settings-bread-service');
 const {generatePrivateSiteAccessCode} = require('./private-site-access-code');
 const {obfuscatedSetting, isSecretSetting, hideValueIfSecret} = require('./settings-utils');
 const mail = require('../mail');
-const {SingleUseTokenProvider} = require('../members/single-use-token-provider');
+const SingleUseTokenProvider = require('../members/single-use-token-provider');
 const urlUtils = require('../../../shared/url-utils');
 
 const ObjectId = require('bson-objectid').default;
@@ -50,8 +50,7 @@ const getSettingsBREADServiceInstance = () => {
             SingleUseTokenModel: models.SingleUseToken,
             validityPeriod: MAGIC_LINK_TOKEN_VALIDITY,
             validityPeriodAfterUsage: MAGIC_LINK_TOKEN_VALIDITY_AFTER_USAGE,
-            maxUsageCount: MAGIC_LINK_TOKEN_MAX_USAGE_COUNT,
-            secret: SettingsCache.get('members_otc_secret')
+            maxUsageCount: MAGIC_LINK_TOKEN_MAX_USAGE_COUNT
         }),
         urlUtils,
         emailAddressService: emailAddressService
